# docs/swagger.yaml
basePath: /
definitions:
...
  handler.CreateSensorConfigRequest:
    properties:
      coef:
        type: number
      id:
        type: string
      index:
        type: integer
      name:
        type: string
      tare:
        type: number
    type: object
...
  handler.PinBuiltinTrainingRequest:
    properties:
      builtin_training_id:
        type: string
    type: object
...
  handler.UpdateSensorConfigRequest:
    properties:
      coef:
        type: number
      index:
        type: integer
      name:
        type: string
      tare:
        type: number
    type: object
...
host: api.crimpy.app
info:
  contact:
    email: support@crimpy.com
    name: API Support
  description: Backend API for Crimpy climbing training application
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://api.crimpy.app/terms
  title: Crimpy API
  version: "1.0"
paths:
...
  /api/pinned-builtin-trainings:
    get:
      consumes:
      - application/json
      description: Retrieve all pinned builtin trainings for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: List of pinned builtin trainings
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all pinned builtin trainings
      tags:
      - PinnedBuiltinTraining
    post:
      consumes:
      - application/json
      description: Pin a builtin training for the authenticated user
      parameters:
      - description: Builtin training to pin
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.PinBuiltinTrainingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Pinned training
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Pin a builtin training
      tags:
      - PinnedBuiltinTraining
  /api/pinned-builtin-trainings/{builtin_training_id}:
    delete:
      consumes:
      - application/json
      description: Remove a pinned builtin training for the authenticated user
      parameters:
      - description: Builtin training ID (UUID)
        in: path
        name: builtin_training_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Unpinned successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Unpin a builtin training
      tags:
      - PinnedBuiltinTraining
  /api/sensor-configs:
    get:
      consumes:
      - application/json
      description: Retrieve all sensor configs for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: List of sensor configs
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all sensor configs
      tags:
      - SensorConfig
    post:
      consumes:
      - application/json
      description: Create a new sensor calibration configuration for the authenticated
        user
      parameters:
      - description: Sensor config details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateSensorConfigRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Sensor config created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a sensor config
      tags:
      - SensorConfig
  /api/sensor-configs/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a sensor config. User must own it.
      parameters:
      - description: Sensor config ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a sensor config
      tags:
      - SensorConfig
    put:
      consumes:
      - application/json
      description: Update a sensor config. User must own it.
      parameters:
      - description: Sensor config ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated sensor config
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateSensorConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated sensor config
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a sensor config
      tags:
      - SensorConfig
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"